Test inherited private properties
Assert that generating accessors for child classes will leave out the
private properties of the parent class.

Currently, there is no use case to generate accessors for the whole
inheritance chain. If the need arises, we'll figure out something.

It might also be a sign of bad design if you need to peek into the
private properties of the parent, so the current behavior is "correct".