Rename project to InspectorGenerator
This improves the domain language because you have "inspectors" that act
on "inspected" objects with private properties.

The previous name, "Spy", was also a misnomer, because a "Spy" is a
special type of test double that implements an interface, storing
internal states passed through the interface methods. The inspector
classes are working with implementations, not interfaces.